Adds a test for explainer
