--- pkg/backend/httpstate/client/test_client.go
@@ -1,30 +1,0 @@
-// Copyright 2025, Pulumi Corporation.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package client
-
-import "net/http"
-
-func NewTestClient(apiURL string, apiToken string, insecure bool, httpClient *http.Client) *Client {
-	return &Client{
-		apiURL:     apiURL,
-		apiToken:   apiAccessToken(apiToken),
-		httpClient: httpClient,
-		restClient: &defaultRESTClient{
-			client: &defaultHTTPClient{
-				client: httpClient,
-			},
-		},
-	}
-}
